Fix Data Type and Duplicate Data
I've fixed data type and duplicate data problem. These problem is causes my iteration to have duplicate copies of data or not iterate at all because of data type problem. I solved it by making the array as an argument to the set method.

- Fixed data type

Fixed duplicate data

Make an array as argument to Set method

See: #4